// src/api/generated/models/TokenPriceMovementDto.ts
/* generated using openapi-typescript-codegen -- do not edit */
/* istanbul ignore file */
/* tslint:disable */
/* eslint-disable */
import type { PriceMovementDto } from './PriceMovementDto';
export type TokenPriceMovementDto = {
    past_24h: PriceMovementDto;
    past_7d: PriceMovementDto;
    past_30d: PriceMovementDto;
    all_time: PriceMovementDto;
};